bugfixes to FileReader
git-svn-id: https://svn.code.sf.net/p/quorum/code/trunk@577 df4a1cb3-768c-44ce-9dd6-df0ee0bb247f

+++ plugins/FilePlugin/src/plugins/quorum/Libraries/System/File.java
@@ -21,7 +21,7 @@ public class File {
     public java.lang.Object $me = null;
     private QuorumFile inst = new QuorumFile();
     
-    public long GetLastModifiedNative() {
+    public double GetLastModifiedNative() {
         return inst.GetLastModifiedNative();
     }
